Make Conque work with t_kb remapped to '\177'
When t_kb is set to '\177' (exe ":set t_kb=\177") Conque imap-s the key
twice and then tries to iunmap it twice too. Second attempt fails as
there is no mapping anymore.

The fix tries to detest the codition and skip mapping character 127. I
was not able to find more elegant way of detecting that t_kb is '\177',
but this should work too.