Do not remap special keys like g:ConqueTerm_EscKey
The keys are mapped in range 33-127. If y ou set g:ConqueTerm_EscKey
to, for example, '+', it's mapping would be overwritten here. So first
detect whether we should or should not map the key.

This is not the only place, earlier in the file there is range 1-31
which will need to be fixed too.